🐞 [GameEngine]: Add demo and game engine bug fixes
A number of improvements and bug fixes are present:

- Remove unused fixed time-step function. Game::Update is now the
  fixed-time step variant, with Game::Render being a variable update
  function. This is to simplify the use of the engine as it may not
  be immediately obvious what fixed_update is for.

- Viewport now uses int32_t for positioning and uint32 for width and
  height instead of floats. This better aligns with OpenGL viewports.

- Changed FPS debug view to use an accumlator over time and to perform
  rounding with use of the lround function for more accurate displaying
  of FPS.

- Window creation happens in the middle of the first active monitor.
  Originally it would always position itself in the top-left, which
  depending on the screen's resolution can be disorientating.

+++ engine/include/Engine/Renderer.hpp
@@ -451,13 +451,13 @@ namespace ASGE {
      * Retrieves the window height.
      * @return The window height.
      */
-    [[nodiscard]] virtual unsigned int windowHeight() const noexcept = 0;
+    [[nodiscard]] virtual float windowHeight() const noexcept = 0;
 
     /**
      * Retrieves the window width.
      * @return The window width.
      */
-    [[nodiscard]] virtual unsigned int windowWidth()  const noexcept = 0;
+    [[nodiscard]] virtual float windowWidth()  const noexcept = 0;
 
     /**
      * @brief Sets the projection matrix
@@ -473,7 +473,7 @@ namespace ASGE {
      * @param[in] width The camera's frustum width.
      * @param[in] height The camera's frustum height.
      */
-    virtual void setProjectionMatrix(float camera_x, float camera_y, float width, float height) = 0;
+    virtual void setProjectionMatrix(int32_t camera_x, int32_t camera_y, uint32_t width, uint32_t height) = 0;
 
     /**
      * @brief Sets the projection matrix
